--- src/components/ingredient/Yuki0kita.md
@@ -1,1 +1,0 @@
-作業しました。
