--- src/components/ingredient/fujishima.md
@@ -1,1 +1,0 @@
-<!-- こんにちは -->
