--- src/components/ingredient/nonn.md
@@ -1,1 +1,0 @@
-作業していました
